refactor(zone-rate): update zone-rate model
