+++ standalone/src/models/shipping-method/zone-rate/types.ts
@@ -1,15 +1,25 @@
-import { Zone, ZoneRate } from '@commercetools/platform-sdk';
+import { ZoneRate, ZoneRateDraft } from '@commercetools/platform-sdk';
 import type { TBuilder } from '@/core';
-import { TCtpZoneRate } from '@/graphql-types';
+import { TCtpZoneRate, TCtpZoneRateDraft } from '@/graphql-types';
 
+export type TZoneRateRest = ZoneRate;
+export type TZoneRateGraphql = TCtpZoneRate;
 /**
  * @deprecated use `TZoneRateRest` or `TZoneRateGraphql` instead
  */
 export type TZoneRate = ZoneRate;
 
-export type TZoneRateRest = ZoneRate;
-export type TZoneRateGraphql = TCtpZoneRate;
+export type TZoneRateDraftRest = ZoneRateDraft;
+export type TZoneRateDraftGraphql = TCtpZoneRateDraft;
+/**
+ * @deprecated use `TZoneRateDraftRest` or `TZoneRateDraftGraphql` instead
+ */
+export type TZoneRateDraft = ZoneRateDraft;
 
 export type TCreateZoneRateBuilder<
-  TZoneRateModel extends TZoneRateRest | TZoneRateGraphql,
+  TZoneRateModel extends
+    | TZoneRateRest
+    | TZoneRateGraphql
+    | TZoneRateDraftRest
+    | TZoneRateDraftGraphql,
 > = () => TBuilder<TZoneRateModel>;
