+++ standalone/src/models/shipping-method/zone-rate/zone-rate-draft/index.ts
@@ -5,8 +5,6 @@ import {
 } from './builders';
 import * as modelPresets from './presets';
 
-export * from './types';
-
 export const ZoneRateDraftRest = {
   random: RestModelBuilder,
   presets: modelPresets.restPresets,
